+++ core/src/main/java/org/eqasim/core/scenario/cutter/population/PopulationCutter.java
@@ -1,9 +1,9 @@
 package org.eqasim.core.scenario.cutter.population;
 
-import java.lang.Thread.UncaughtExceptionHandler;
 import java.util.Iterator;
 import java.util.LinkedList;
 import java.util.List;
+import java.util.concurrent.atomic.AtomicBoolean;
 
 import org.eqasim.core.misc.Constants;
 import org.eqasim.core.misc.ParallelProgress;
@@ -39,9 +39,14 @@ public class PopulationCutter {
 		ParallelProgress progress = new ParallelProgress("Cutting population ...", population.getPersons().size());
 		progress.start();
 
+		AtomicBoolean errorsOccured = new AtomicBoolean(false);
+
 		for (int i = 0; i < numberOfThreads; i++) {
 			Thread thread = new Thread(new Worker(personIterator, progress, planCutterProvider));
-			thread.setUncaughtExceptionHandler(new ExceptionHandler());
+			thread.setUncaughtExceptionHandler((t, e) -> {
+				e.printStackTrace();
+				errorsOccured.set(true);
+			});
 
 			thread.start();
 			threads.add(thread);
@@ -52,13 +57,9 @@ public class PopulationCutter {
 		}
 
 		progress.close();
-	}
 
-	private class ExceptionHandler implements UncaughtExceptionHandler {
-		@Override
-		public void uncaughtException(Thread t, Throwable e) {
-			e.printStackTrace();
-			throw new RuntimeException(e);
+		if (errorsOccured.get()) {
+			throw new RuntimeException("Errors occured while cutting the population.");
 		}
 	}